+++ libs/movex-server/src/lib/movex-server.ts
@@ -107,7 +107,6 @@ export const movexServer = <TDefinition extends MovexDefinition>(
       {
         id: clientId,
         info: clientInfo,
-        clockOffset: 0, // start it at 0
       }
     );
 
@@ -115,24 +114,10 @@ export const movexServer = <TDefinition extends MovexDefinition>(
 
     movexMaster.addClientConnection(connectionToClient);
 
-    // socket.on('pong', (p) => {
-    //   console.log('on pong', p);
-    // });
-
-    // io.on('pong', (p) => {
-    //   console.log('on pong', p);
-    // });
-
-    // const pingIntervalId = setInterval(() => {
-    //   io.emit('ping', { clientId });
-    // }, 10 * 1000);
-
     io.on('disconnect', () => {
       logsy.info('Client Disconnected', { clientId });
 
       movexMaster.removeConnection(clientId);
-
-      // clearInterval(pingIntervalId);
     });
   });
 
@@ -185,12 +170,7 @@ export const movexServer = <TDefinition extends MovexDefinition>(
 
   // Public State
   app.get('/api/resources/:rid/state', async (req, res) => {
-    const masterContext = createMasterContext({
-      extra: {
-        clientId: 'UNKNOWN',
-        req: 'getPublicResourceCheckedState',
-      },
-    });
+    const masterContext = createMasterContext();
 
     const rawRid = req.params.rid;
 
@@ -234,7 +214,7 @@ export const movexServer = <TDefinition extends MovexDefinition>(
   //   }
   // });
 
-  // //start our server
+  // start the server
   const port = process.env['port'] || 3333;
   httpServer.listen(port, () => {
     const address = httpServer.address();